requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
distlib==0.3.6
emoji==2.1.0
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.1
google-search-results==2.4.1
greenlet==2.0.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
imageio==2.26.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.1
matplotlib==3.6.1
multidict==6.0.2
networkx==3.0
numpy==1.23.4
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.2
Pillow==9.2.0
platformdirs==2.5.2
pydicom==2.3.1
pyee==9.0.4
pylidc==0.2.3
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyWavelets==1.4.1
requests==2.28.1
rfc3986==1.5.0
scikit-image==0.20.0
scipy==1.10.1
selenium==4.5.0
six==1.16.0
slack==0.0.2
slack-bolt==1.15.2
slack-sdk==3.19.1
slackeventsapi==3.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.7
tifffile==2023.3.21
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.6
webdriver-manager==3.8.4
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.1
